% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmatrix.R
\name{year_cormatrix}
\alias{year_cormatrix}
\title{A Basketball Function of a Correlation Matrix}
\usage{
year_cormatrix(x)
}
\arguments{
\item{cormatrix}{Returns the correlation matrix for that year}
}
\description{
This function allows you to input a year and find out the correlation matrix of all numeric vairables in that year.
}
\examples{
year_cormatrix()
}
\keyword{basketball,}
\keyword{correlation,}
\keyword{matrix}
\keyword{year,}
